update use reduce and fix bug

+++ Backend/app/Services/JobSeekerService.php
@@ -46,7 +46,10 @@ class JobSeekerService
         if ($jobSeeker) {
             $old_path = $jobSeeker->logo;
             if (isset($old_path) and Storage::disk('public')->exists($old_path)) {
-                Storage::disk('public')->delete($old_path);
+                if($old_path!="images/userlogo.png"){
+                    Storage::disk('public')->delete($old_path);
+
+                }
             }
             $this->jobSeekerRepository->saveLogoPath($jobSeeker, $path);
         }
